Split tidy option page in tidy options and tidy check
Created separate options pages for tidy options and tidy check.
Move all option pages in DialogPages folder.
