+++ ClangPowerTools/ClangPowerTools/DialogPages/GeneralOptions.cs
