+++ ClangPowerTools/ClangPowerTools/DialogPages/TidyChecks.cs
@@ -3,53 +3,9 @@ using System.ComponentModel;
 
 namespace ClangPowerTools
 {
-  public class TidyOptions : DialogPage
+  public class TidyChecks : DialogPage
   {
-    #region Members
-
-    private string[] mTidyChecks;
-
-    #endregion
-
-    #region Methods
-
-    public override void LoadSettingsFromStorage()
-    {
-      base.LoadSettingsFromStorage();
-      TidyChecks = UseUserSettings ? mTidyChecks : DefaultOptions.kTidyChecks;
-      UseUserSettings = true;
-    }
-
-    #endregion
-
     #region Properties
-
-    #region Tidy Category
-
-    [Category(" Tidy")]
-    [DisplayName("Custom Checks")]
-    [Description("If not empty clang-tidy will be called with given flags, instead of clang++. The tidy operation is applied to whole translation units, meaning all directory headers included in the CPP will be tidied up too. Changes will not be applied, only simulated.")]
-    [TypeConverter(typeof(StringArrayConverter))]
-    public string[] TidyChecks
-    {
-      get => mTidyChecks;
-      set => mTidyChecks = value;
-    }
-
-    [Category(" Tidy")]
-    [DisplayName("Fix")]
-    [Description("If not empty clang-tidy will be called with given flags, instead of clang++. The tidy operation is applied to whole translation units, meaning all directory headers included in the CPP will be tidied up too. Changes will be applied to the file(s).")]
-    public bool Fix { get; set; }
-
-    [Category(" Tidy")]
-    [DisplayName("Use User Settings")]
-    [Description("If not empty clang-tidy will be called with given flags, instead of clang++. The tidy operation is applied to whole translation units, meaning all directory headers included in the CPP will be tidied up too. Changes will not be applied, only simulated.")]
-    [Browsable(false)]
-    public bool UseUserSettings { get; set; }
-
-    #endregion
-
-    #region Checks Category
 
     [Category("Checks")]
     [DisplayName("clang-analyzer-core.CallAndMessage")]
@@ -1554,8 +1510,6 @@ namespace ClangPowerTools
     [DisplayName("readability-implicit-bool-cast")]
     [ClangCheck(true)]
     public bool ReadabilityImplicitBoolCast { get; set; } = true;
-
-    #endregion
 
     #endregion
