+++ ClangPowerTools/ClangPowerTools/DialogPages/TidyPage.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.VisualStudio.Shell;
+
+namespace ClangPowerTools
+{
+  public class TidyPage : DialogPage
+  {
+  }
+}
